Extract static content handling into reusable class
